+++ deps/cli.ts
@@ -4,4 +4,3 @@ export * from "./common.ts";
 
 export * as cliffy_cmd from "https://deno.land/x/cliffy@v1.0.0-rc.4/command/mod.ts";
 export { Table } from "https://deno.land/x/cliffy@v1.0.0-rc.4/table/table.ts";
-export * as zod_val_err from "npm:zod-validation-error@3.3.0";
