+++ host/mod.ts
@@ -20,14 +20,23 @@ export interface CliArgs {
   ghjkfilePath?: string;
   ghjkDirPath?: string;
   reFlagSet: boolean;
+  lockedFlagSet: boolean;
 }
 
 type HostCtx = {
   fileHashMemoStore: Map<string, Promise<string>>;
   curEnvVars: Record<string, string>;
+  reFlagSet: boolean;
+  lockedFlagSet: boolean;
 };
+// FIXME: better way to resolve version
+const GHJK_VERSION = "0.2.0";
 
 export async function cli(args: CliArgs) {
+  logger().debug(`ghjk CLI`, GHJK_VERSION);
+  if (args.reFlagSet && args.lockedFlagSet) {
+    throw new Error("GHJK_LOCKED && GHJK_RE both set");
+  }
   // items to run at end of function
   const defer = [] as (() => Promise<any>)[];
 
@@ -52,7 +61,12 @@ export async function cli(args: CliArgs) {
         ? $.path(args.ghjkfilePath).resolve().normalize()
         : undefined,
       blackboard: new Map(),
+    };
+    const hcx: HostCtx = {
+      fileHashMemoStore: new Map(),
+      curEnvVars: Deno.env.toObject(),
       reFlagSet: args.reFlagSet,
+      lockedFlagSet: args.lockedFlagSet,
     };
     logger().debug("context established", {
       ghjkDir: gcx?.ghjkDir.toString(),
@@ -67,7 +81,7 @@ export async function cli(args: CliArgs) {
 
     // this returns nothing if no valid lockifle or ghjkfile
     // is found
-    const commands = await commandsFromConfig(gcx);
+    const commands = await commandsFromConfig(hcx, gcx);
     if (commands) {
       serializedConfig = commands.config;
       // lock entries are also generated across program usage
@@ -90,7 +104,7 @@ export async function cli(args: CliArgs) {
 
   const root = new cliffy_cmd.Command()
     .name("ghjk")
-    .version("0.1.1") // FIXME: better way to resolve version
+    .version(GHJK_VERSION)
     .description("Programmable runtime manager.")
     .action(function () {
       this.showHelp();
@@ -183,17 +197,21 @@ export async function cli(args: CliArgs) {
   await Promise.all(defer.map((fn) => fn()));
 }
 
-async function commandsFromConfig(gcx: GhjkCtx) {
-  const hcx: HostCtx = {
-    fileHashMemoStore: new Map(),
-    curEnvVars: Deno.env.toObject(),
-  };
-
+async function commandsFromConfig(hcx: HostCtx, gcx: GhjkCtx) {
   const lockFilePath = gcx.ghjkDir.join("lock.json");
   const hashFilePath = gcx.ghjkDir.join("hash.json");
 
   const foundLockObj = await readLockFile(lockFilePath);
   const foundHashObj = await readHashFile(hashFilePath);
+
+  if (hcx.lockedFlagSet) {
+    if (!foundLockObj) {
+      throw new Error("GHJK_LOCKED set but no lockfile found");
+    }
+    if (!foundHashObj) {
+      throw new Error("GHJK_LOCKED set but no hashfile found");
+    }
+  }
 
   const lockEntries = {} as Record<string, unknown>;
 
@@ -201,7 +219,7 @@ async function commandsFromConfig(gcx: GhjkCtx) {
     ? await fileDigestHex(hcx, gcx.ghjkfilePath!)
     : undefined;
 
-  if (!gcx.reFlagSet && foundLockObj) {
+  if (!hcx.reFlagSet && foundLockObj) {
     logger().debug("loading lockfile", lockFilePath);
     for (const man of foundLockObj.config.modules) {
       const mod = std_modules.map[man.id];
@@ -223,16 +241,19 @@ async function commandsFromConfig(gcx: GhjkCtx) {
       );
     }
   }
+
   let configExt: SerializedConfigExt | null = null;
+  let wasReSeraialized = false;
   if (
-    !gcx.reFlagSet &&
+    !hcx.reFlagSet &&
     foundLockObj &&
     foundHashObj &&
-    foundLockObj.version == "0" &&
-    // avoid reserializing the config if
-    // the ghjkfile and environment is _satisfcatorily_
-    // similar. "cache validation"
-    await isHashFileValid(hcx, foundLockObj, foundHashObj, ghjkfileHash)
+    (hcx.lockedFlagSet ||
+      // avoid reserializing the config if
+      // the ghjkfile and environment is _satisfcatorily_
+      // similar. "cache validation"
+      foundLockObj.version == "0" &&
+        await isHashFileValid(hcx, foundLockObj, foundHashObj, ghjkfileHash))
   ) {
     configExt = {
       config: foundLockObj.config,
@@ -243,6 +264,7 @@ async function commandsFromConfig(gcx: GhjkCtx) {
   } else if (gcx.ghjkfilePath) {
     logger().info("serializing ghjkfile", gcx.ghjkfilePath);
     configExt = await readGhjkfile(hcx, gcx.ghjkfilePath);
+    wasReSeraialized = true;
   } else {
     // nothing to get the commands from
     return;
@@ -284,7 +306,12 @@ async function commandsFromConfig(gcx: GhjkCtx) {
       subCommands[cmdName] = [cmd, man.id];
     }
   }
-  if (!foundHashObj || !deep_eql(newHashObj, foundHashObj)) {
+
+  if (
+    !hcx.lockedFlagSet && wasReSeraialized && (
+      !foundHashObj || !deep_eql(newHashObj, foundHashObj)
+    )
+  ) {
     await hashFilePath.writeJsonPretty(newHashObj);
   }
 
@@ -298,6 +325,8 @@ async function commandsFromConfig(gcx: GhjkCtx) {
     subCommands,
     config: configExt.config,
     async writeLockFile() {
+      if (hcx.lockedFlagSet) return;
+
       const newLockObj: zod.infer<typeof lockObjValidator> = {
         version: "0",
         platform: serializePlatform(Deno.build),
