+++ main.ts
@@ -26,6 +26,8 @@ if (import.meta.main) {
     // FIXME: better
     reFlagSet: !!Deno.env.get("GHJK_RE") &&
       !(["false", "", ""].includes(Deno.env.get("GHJK_RE")!)),
+    lockedFlagSet: !!Deno.env.get("GHJK_LOCKED") &&
+      !(["false", "", ""].includes(Deno.env.get("GHJK_RE")!)),
 
     ghjkShareDir: Deno.env.get("GHJK_SHARE_DIR") ??
       dirs().shareDir.resolve("ghjk").toString(),
