+++ modules/types.ts
@@ -18,7 +18,6 @@ export type GhjkCtx = {
   ghjkDir: Path;
   ghjkShareDir: Path;
   blackboard: Map<string, unknown>;
-  reFlagSet: boolean;
 };
 
 export default {
