+++ tests/utils.ts
@@ -193,10 +193,10 @@ export function genTsGhjkFile(
     2,
   );
 
-  const tasks = (secureConf?.tasks ?? []).map(
-    (def) => {
+  const tasks = Object.entries(secureConf?.tasks ?? {}).map(
+    ([name, def]) => {
       const stringifiedSection = JSON.stringify(
-        def,
+        { ...def, name },
         (_, val) =>
           typeof val == "string" ? val.replaceAll(/\\/g, "\\\\") : val,
         2,
